+++ packages/base-model/test/manuscript_test.js
@@ -11,6 +11,13 @@ describe('Manuscript', () => {
     await dbCleaner()
   })
 
+  it('has upated set when created', async () => {
+    const manuscript = await new Manuscript({ title: 'Test' }).save()
+    expect(manuscript.title).toEqual('Test')
+    const now = new Date().toISOString()
+    expect(manuscript.updated).toHaveLength(now.length)
+  })
+
   it('can be saved and found and deleted', async () => {
     const manuscript = await new Manuscript({ title: 'Test' }).save()
     expect(manuscript.title).toEqual('Test')
@@ -79,5 +86,25 @@ describe('Manuscript', () => {
     const id = '1838d074-fb9d-4ed6-9c63-39e6bc7429ce'
     const manuscript = await new Manuscript({ id }).save()
     expect(manuscript.id).toEqual(id)
+  })
+
+  it('old data does not overwrite new', async () => {
+    // T0 - start time (A == B)
+    let manuscriptA = await new Manuscript({ title: 'T0' }).save()
+    expect(manuscriptA.title).toEqual('T0')
+    const manuscriptB = await Manuscript.find(manuscriptA.id)
+
+    // T1 - B is changed (not saved)
+    manuscriptB.title = 'T1'
+
+    // T2 - A is changed and saved
+    manuscriptA.title = 'T2'
+    manuscriptA = await manuscriptA.save()
+    expect(manuscriptA.updated).not.toBe(manuscriptB.updated)
+
+    // T4 - now save B, this should throw as `updated` is older than current.
+    await expect(manuscriptB.save()).rejects.toThrow(
+      'Data Integrity Error property updated',
+    )
   })
 })