+++ packages/components/FormGroup/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.32](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-form-group@1.1.31...pubsweet-component-form-group@1.1.32) (2018-12-04)
+
+**Note:** Version bump only for package pubsweet-component-form-group
+
+
+
+
+
+## [1.1.31](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-form-group@1.1.30...pubsweet-component-form-group@1.1.31) (2018-11-30)
+
+**Note:** Version bump only for package pubsweet-component-form-group
+
+
+
+
+
+## [1.1.30](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-form-group@1.1.29...pubsweet-component-form-group@1.1.30) (2018-11-29)
+
+**Note:** Version bump only for package pubsweet-component-form-group
+
+
+
+
+
 <a name="1.1.29"></a>
 ## [1.1.29](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-form-group@1.1.28...pubsweet-component-form-group@1.1.29) (2018-11-13)