+++ packages/components/xpub-manuscript/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.3](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-manuscript@0.6.2...pubsweet-component-xpub-manuscript@0.6.3) (2018-12-04)
+
+**Note:** Version bump only for package pubsweet-component-xpub-manuscript
+
+
+
+
+
+## [0.6.2](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-manuscript@0.6.1...pubsweet-component-xpub-manuscript@0.6.2) (2018-11-30)
+
+**Note:** Version bump only for package pubsweet-component-xpub-manuscript
+
+
+
+
+
+## [0.6.1](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-manuscript@0.6.0...pubsweet-component-xpub-manuscript@0.6.1) (2018-11-29)
+
+**Note:** Version bump only for package pubsweet-component-xpub-manuscript
+
+
+
+
+
 <a name="0.6.0"></a>
 # [0.6.0](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-manuscript@0.5.7...pubsweet-component-xpub-manuscript@0.6.0) (2018-11-05)